Moved App.jsx logic to Boards.jsx
